feat: add phase 2 conversation models
