+++ conversation_service/models/__init__.py
@@ -1,0 +1,21 @@
+from .conversation_models import (
+    ConversationRequest,
+    ConversationResponse,
+    ConversationMetadata,
+    ConversationContext,
+)
+from .conversation_db_models import (
+    Conversation,
+    ConversationSummary,
+    ConversationTurn,
+)
+
+__all__ = [
+    "ConversationRequest",
+    "ConversationResponse",
+    "ConversationMetadata",
+    "ConversationContext",
+    "Conversation",
+    "ConversationSummary",
+    "ConversationTurn",
+]
